requirements.txt
playwright~=1.47
python-dotenv~=1.0
